pyproject.toml
[project]
name = "medlive"
version = "0.1.0"
description = "MedLive - Real-time multilingual AI health guardian powered by Gemini Live API and Google ADK"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.2.0",
    "google-genai>=1.14.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-cloud-firestore>=2.19.0",
    "firebase-admin>=6.6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "twilio>=9.0.0",
    "httpx>=0.27.0",
    "google-cloud-tasks>=2.21.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
